refactor(api): move usecase
update-user-for-account-recovery
